Fix the MoD, Done.

+++ source/frost/transferlayer/MessageDownloader.java
@@ -88,6 +88,14 @@ public class MessageDownloader {
             return null;
         }
         
+ 	    if(results != null && results.getDoneBlocks() > 0) {
+ 	    	logger.severe("TOFDN: Contents of message key partially missing."+logInfo);
+ 	    	System.out.println("TOFDN: Contents of message key partially missing.");
+ 	    	MessageDownloaderResult mdResult = new MessageDownloaderResult();
+ 	    	mdResult.errorMsg = MessageDownloaderResult.BROKEN_KSK;
+	    	return mdResult;
+ 	    }
+        
         if( results == null || results.isSuccess() == false ) {
             tmpFile.delete();
             return null;
